requirements.txt
bokeh==1.1.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
gitdb2==2.0.5
GitPython==2.1.11
h5py==2.9.0
idna==2.8
Jinja2==2.10.1
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
networkx==2.3
numpy==1.16.3
packaging==19.0
pandas==0.24.2
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
smmap2==2.0.5
torch==1.0.1.post2
torchfile==0.1.0
torchvision==0.2.2.post3
tornado==6.0.2
tqdm==4.31.1
urllib3==1.24.2
visdom==0.1.8.8
websocket-client==0.56.0
